requirements.txt
google-generativeai>=0.3.0
prompt-toolkit>=3.0.0
pathlib2>=2.3.0; python_version < "3.4" 